新增心跳功能避免浏览器强退导致redis键值没有删除进而导致锁死的bug，修复了多端登录日志重复插入的bug

+++ RuoY/ruoyi-framework/src/main/java/com/ruoyi/framework/web/service/TokenService.java
@@ -41,7 +41,7 @@ public class TokenService
     @Value("${token.secret}")
     private String secret;
 
-    // 令牌有效期（默认30分钟）
+    // 令牌有效期（默认60分钟）
     @Value("${token.expireTime}")
     private int expireTime;
 
@@ -101,7 +101,12 @@ public class TokenService
         if (StringUtils.isNotEmpty(token))
         {
             String userKey = getTokenKey(token);
-            redisCache.deleteObject(userKey);
+            LoginUser loginUser = redisCache.getCacheObject(userKey);
+            if (StringUtils.isNotNull(loginUser))
+            {
+                redisCache.deleteObject(userKey);
+                redisCache.deleteObject(CacheConstants.LOGIN_USER_TOKEN + loginUser.getUsername());
+            }
         }
     }
